Add bundle capabilities, Tycho pom.xml and javax.usb.api
Combining org.usb4java.javax and javax.usb.api in the same plugin/jar is
a workaround to make the javax.usb.api and the javax.usb.properties file
work in an Eclipse plug-in context. This probably should and can be done
differently. For now it's a shortcut after wasting multiple nights
dealing with this dependency hell.